Move static scope into it's own file
